uop: move xpow to symbolic, breaking symbolic-codegen cycle
